Added methods and getters to EdgeBranch.
Added flag to test if it is an EdgeBranch.
Need to decide whether static branch updater should raise an exception if the resolved branch is not a branched edge - I think so…
Need to insert edge if it doesn’t exist in static method: need to refactor method flag parameters.