Revised Document family.
Added restrictedFields() getter: returns list of fields that the document must not have.
Reviewed the logiv and should be final now.
Studied transactions: should add a transactions folder with functions that execute a transaction: the database will handle this.
Studied tests, will implement tests for Document family as a starter test of test.
Good night!