Finalised test replace values.
Should refactor title to reflect better what is going on: We load the existing document in the class instance, we modify the values of the document directly in the database and we replace the document.
Need to replicate this test where we load the existing document in the class instance, we update the fields directly without going through setDocumentProperties() and we replace the document: this should catch errors in the validation.